Fix UnboundLocalError in error case
When the test-process cannot be started, the finally block errors out
with an UnboundLocalError.

Example:
```
          finally:
  >           p.kill()
  E           UnboundLocalError: local variable 'p' referenced before assignment

  tests/__init__.py:62: UnboundLocalError
```

+++ notebooks/tests/__init__.py
@@ -53,13 +53,15 @@ def _remove_folders(input_folders: [str]) -> None:
 def start_nessie() -> subprocess.Popen:
     """Context for starting and stopping a nessie binary."""
     start_command = _fetch_and_get_nessie_start_command()
+    p = None
     try:
         p = subprocess.Popen(  # noqa: S603
             start_command, stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL
         )
         yield p
     finally:
-        p.kill()
+        if p:
+            p.kill()
 
 
 def _fetch_and_get_nessie_start_command() -> [str]:
